Use SRV records as service discovery for IDS
With some reliability issues with our upstream PaaS, we're seeing
mysterious issues on connect.

This PR adds a level of indirection to allow us to point I.D.S. to
a variety of hosts using SRV records.

SRV records look like this:

                                                  priority   weight
                                                           \  |  port   hostname
    _detsys_ids._tcp.install.determinate.systems. 60 IN SRV 1 10 443 A.install.determinate.systems.
    _detsys_ids._tcp.install.determinate.systems. 60 IN SRV 1 10 443 B.install.determinate.systems.

* Priority: 1 > 10, records of high (low numerically) priority are
  chosen first.
* Multiple records with the same priority are chosen based on the weight
  where a larger numerical weight value makes it more likely to be
  picked.

Then the client picks a record, and connects to its hostname on the
defined port. This allows us to have multiple providers with independent
domains and TLS certificates.

NOTE: This PR changes the behavior around diagnostics URLs, which is
represented by the type change of the ActionOptions. Instead of passing
the full diagnostics URL, projects provide the URL suffix. This means
that actions need to remove the default value for `diagnostics-endpoint`,
and make it a non-required option.